+++ src/view/Matcher/Verbose.js
